refactor(donate): abstract the layout as frequency-based layout (#1284)
