+++ package.json
@@ -11,7 +11,7 @@
 		"html-entities": "^2.3.3",
 		"identity-obj-proxy": "^3.0.0",
 		"lint-staged": "^13.0.3",
-		"newspack-scripts": "^4.3.0",
+		"newspack-scripts": "^4.3.1",
 		"postcss-scss": "^4.0.5",
 		"prettier": "npm:wp-prettier@^2.6.2-beta-1",
 		"stylelint": "^14.9.1"
