+++ src/blocks/donate/frequency-based/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Internal dependencies
+ */
+import './style.scss';
