+++ src/blocks/donate/frequency-based/style.scss
@@ -1,8 +1,8 @@
-@use '../../shared/sass/colors';
-@use '../../shared/sass/mixins';
-@use '../../shared/sass/variables';
+@use '../../../shared/sass/colors';
+@use '../../../shared/sass/mixins';
+@use '../../../shared/sass/variables';
 
-.wpbnbd {
+.wpbnbd--frequency-based {
 	background: colors.$color__background-body;
 	border: 1px solid variables.$color__border;
 	color: colors.$color__text-main;
@@ -265,238 +265,5 @@
 				top: 0;
 			}
 		}
-	}
-}
-
-/* Alternate Style */
-
-.wpbnbd.is-style-alternate {
-	border-color: colors.$color__background-screen;
-	border-radius: 5px;
-
-	.thanks {
-		margin: 1.12rem;
-	}
-
-	.money-input {
-		max-width: 100%;
-
-		input[type='number'] {
-			max-width: 100%;
-			width: 100%;
-		}
-	}
-
-	&.untiered {
-		.input-container {
-			margin: 1.12rem;
-		}
-	}
-
-	&.tiered {
-		.tiers {
-			.tier-label {
-				background: colors.$color__background-screen;
-				border-color: colors.$color__background-screen;
-				margin: 0;
-				text-align: center;
-			}
-		}
-
-		.money-input {
-			bottom: 0;
-			left: 0;
-			right: 0;
-		}
-	}
-
-	.frequencies {
-		@include mixins.media( tablet ) {
-			padding-top: calc( 1.14rem + 1.6em + 4px );
-		}
-	}
-
-	.frequency {
-		.freq-label {
-			background: colors.$color__background-screen;
-			border-color: colors.$color__background-screen;
-			border-width: 1px;
-			padding: calc( 0.57rem + 1px ) 0.5rem;
-			@include mixins.media( mobile ) {
-				padding: calc( 0.57rem + 1px ) 0.76rem;
-			}
-		}
-
-		input[type='radio']:checked {
-			+ .freq-label {
-				&::after {
-					box-sizing: content-box;
-					background: white;
-					border-radius: 5px;
-					bottom: 0.19rem;
-					border: 0.19rem solid white;
-					box-shadow: 0 0 0 1px variables.$color__border;
-					content: '';
-					display: block;
-					left: 0.19rem;
-					position: absolute;
-					right: 0.19rem;
-					top: 0.19rem;
-					z-index: -1;
-				}
-
-				&:hover {
-					background: colors.$color__background-screen;
-				}
-			}
-
-			~ .tiers {
-				display: grid;
-				grid-gap: 0.28rem;
-				grid-template-columns: repeat( 2, 1fr );
-				margin: 1.12rem;
-
-				@include mixins.media( mobile ) {
-					grid-template-columns: repeat( 4, 1fr );
-					grid-gap: 0.56rem;
-				}
-			}
-		}
-	}
-
-	button {
-		border-top-left-radius: 0;
-		border-top-right-radius: 0;
-		margin: 0 -1px -1px;
-		width: calc( 100% + 2px );
-	}
-}
-
-/* Minimal Style */
-
-.wpbnbd.is-style-minimal {
-	border: none;
-
-	form {
-		text-align: center;
-	}
-
-	.thanks {
-		margin: 1.12rem 0;
-	}
-
-	.money-input {
-		max-width: 100%;
-
-		input[type='number'] {
-			max-width: 100%;
-			width: 100%;
-		}
-	}
-
-	&.untiered {
-		.input-container {
-			margin: 0.56rem 0 1.12rem;
-			text-align: left;
-		}
-	}
-
-	&.tiered {
-		.tiers {
-			.tier-label {
-				align-items: center;
-				display: grid;
-				background: transparent;
-				border: none;
-				color: colors.$color__text-light;
-				font-family: inherit;
-				font-weight: inherit;
-				grid-gap: 0.25rem;
-				grid-template-columns: auto 1fr;
-				margin: 0;
-				padding: 0;
-				text-align: left;
-
-				&::before {
-					background: white;
-					border: 1px solid variables.$color__border;
-					border-radius: 100%;
-					box-shadow: inset 0 0 0 3px white;
-					content: '';
-					display: block;
-					height: 20px;
-					padding: 3px;
-					width: 20px;
-				}
-
-				&:hover {
-					color: inherit;
-				}
-			}
-
-			input[type='radio']:checked + .tier-select-label {
-				background: none;
-				color: inherit;
-
-				&::before {
-					background: currentcolor;
-				}
-			}
-		}
-
-		.money-input {
-			bottom: 0;
-			left: 0;
-			right: 0;
-		}
-	}
-
-	.frequencies {
-		padding-top: calc( 0.76rem + 1.6em + 4px );
-	}
-
-	.frequency {
-		.freq-label {
-			background: transparent;
-			border: none;
-			box-shadow: 0 3px 0 white, 0 4px 0 0 variables.$color__border;
-			padding: 0.38rem 0;
-			transition: none;
-
-			@include mixins.media( tablet ) {
-				padding-left: 0.76rem;
-				padding-right: 0.76rem;
-			}
-
-			&::before {
-				border-width: 1px;
-				border-color: variables.$color__border;
-			}
-		}
-
-		input[type='radio']:checked {
-			+ .freq-label {
-				box-shadow: 0 4px 0 currentcolor;
-			}
-
-			~ .tiers {
-				display: flex;
-				flex-wrap: wrap;
-				grid-gap: 1.12rem;
-				justify-content: center;
-				margin: 1.12rem 0;
-
-				@include mixins.media( mobile ) {
-					display: grid;
-					grid-gap: 0.56rem;
-					grid-template-columns: repeat( 4, 1fr );
-				}
-			}
-		}
-	}
-
-	button {
-		margin: 0 auto;
-		min-width: 50%;
 	}
 }
