+++ src/blocks/donate/streamlined/index.ts
@@ -15,7 +15,6 @@ import 'regenerator-runtime'; // Required in WP >=5.8.
  */
 import * as utils from './utils';
 import './style.scss';
-import type { DonationSettings } from '../types';
 
 export const processStreamlinedElements = ( parentElement = document ) =>
 	[ ...parentElement.querySelectorAll( '.stripe-payment' ) ].forEach( async el => {
@@ -314,9 +313,7 @@ export const processStreamlinedElements = ( parentElement = document ) =>
 			);
 
 			const formValues = utils.getDonationFormValues( formElement );
-			const validationErrors = Object.values(
-				utils.validateFormData( formValues, settings as DonationSettings )
-			);
+			const validationErrors = Object.values( utils.validateFormData( formValues, settings ) );
 			if ( validationErrors.length > 0 ) {
 				utils.renderMessages( validationErrors, messagesEl );
 				enableForm();
