+++ src/blocks/donate/streamlined/utils.ts
@@ -10,13 +10,16 @@ import type { DonationFormValues, DonationFormInputName, DonationSettings } from
 
 const isValidEmail = ( string: string ) => /\S+@\S+/.test( string );
 
-export const validateFormData = ( values: DonationFormValues, settings: DonationSettings ) => {
+export const validateFormData = (
+	values: DonationFormValues,
+	settings: Partial< DonationSettings >
+) => {
 	const errors: { [ key: string ]: string } = {};
 	if ( ! isValidEmail( values.email ) ) {
 		errors.email = __( 'Email address is invalid.', 'newspack-blocks' );
 	}
 	const { minimumDonation } = settings;
-	if ( parseFloat( values.amount ) < minimumDonation ) {
+	if ( minimumDonation && parseFloat( values.amount ) < minimumDonation ) {
 		errors.amount = sprintf(
 			/* Translators: %d is minimum donation amount set in Reader Revenue wizard or block attributes. */
 			__( 'Amount must be at least %d.', 'newspack-blocks' ),
