+++ src/blocks/donate/styles/view.scss
@@ -1,0 +1,1 @@
+@import './style-variations';
